Привёл в порядок, разбил большие классы на отдельные файлы, удалил неиспользуемые импорты
