+++ src/event/edit/final_edit.py
@@ -1,0 +1,19 @@
+from src.database.db_draft_operations import delete_draft
+
+
+async def finalize_edit(context, draft):
+    """Завершает редактирование"""
+    from src.message.send_message import send_event_message
+
+    # Обновляем сообщение мероприятия
+    await send_event_message(
+        event_id=draft["event_id"],
+        context=context,
+        chat_id=draft["chat_id"],
+        message_id=draft["original_message_id"]
+    )
+
+    # Удаляем черновик
+    delete_draft(context.bot_data["drafts_db_path"], draft["id"])
+
+
