+++ src/handlers/mention_handler.py
@@ -3,11 +3,10 @@ from telegram.ext import ContextTypes, MessageHandler, filters
 from telegram.error import BadRequest
 
 from src.database.db_draft_operations import add_draft, get_user_chat_draft, update_draft
-from src.handlers.draft_handlers import handle_draft_message
 from src.logger.logger import logger
 
 async def mention_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    # Логируем входящее сообщение для отладки
+    # Отправляем лог о входящем сообщении для отладки
     logger.debug(f"Получено сообщение: {update.message.text}")
 
     if not update.message or not update.message.entities:
